REF: consolidate __iter__ into TimelikeOps, standardize _validate_setitem_value naming
Move the chunked __iter__ implementation shared by DatetimeArray and
TimedeltaArray into their common base class TimelikeOps, with each
subclass providing only a _iter_convert_chunk hook.

Rename _maybe_convert_setitem_value to _validate_setitem_value in
ArrowExtensionArray, StringArray, and ArrowStringArray so all array
types use the same interface name for setitem validation.
